+++ config/initializers/nibble_content.rb
@@ -1,0 +1,1 @@
+Rails.application.config.to_prepare { Nibble::Files.watch! } if Rails.env.development?